+++ CHANGELOG.md
@@ -7,8 +7,9 @@ dd mm yyyy
 - Change: Apps with the `app-esi` role no longer automatically have access to the default ESI token. Instead, 
   the corresponding EVE login must now be added to the app (this is done automatically for all existing app
   with a database migration).
+- Change: Added `/api/app/v1/esi/eve-login/{name}/characters` now allows the "core.default" login.
 - App API: Implemented EVE login permission check for apps (ESI endpoint) ([#32][i32]).
-- App API: Added `/app/v1/esi/eve-login/{name}/valid-token-data` endpoint.
+- App API: Added `/api/app/v1/esi/eve-login/{name}/token-data` endpoint.
 
 ## 1.35.0